+++ src/Microsoft.AspNet.Hosting/Startup/ApplicationStartup.cs
@@ -2,7 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.Collections.Generic;
 using System.Globalization;
+using System.Linq;
 using System.Reflection;
 using Microsoft.AspNet.Builder;
 using Microsoft.Framework.DependencyInjection;
@@ -15,7 +17,7 @@ namespace Microsoft.AspNet.Hosting.Startup
         {
             Instance = startupInstance;
             ConfigureMethod = configure;
-            ConfigureServicesMethod = ConfigureServicesMethod;
+            ConfigureServicesMethod = configureServices;
         }
 
         public ConfigureServicesDelegate ConfigureServicesMethod { get; }
@@ -23,7 +25,7 @@ namespace Microsoft.AspNet.Hosting.Startup
 
         public object Instance { get; }
 
-        // REVIEW: is serviceProvider always builder.ApplicationServices here??
+        // REVIEW: need to revisit the import implications here (are services the exported services?)
         public IServiceProvider ConfigureServices(IApplicationBuilder builder, IServiceCollection services)
         {
             return ConfigureServicesMethod == null
@@ -31,11 +33,73 @@ namespace Microsoft.AspNet.Hosting.Startup
                 : ConfigureServicesMethod.Invoke(Instance, builder, services);
         }
 
-        // REVIEW: is serviceProvider always builder.ApplicationServices here??
-        public void Configure(IApplicationBuilder builder)
+        public void Configure(IServiceProvider services, IApplicationBuilder builder)
         {
-            ConfigureMethod.Invoke(Instance, builder);
+            ConfigureMethod.Invoke(Instance, services, builder);
         }
+
+        public static ApplicationStartup LoadStartup(
+            IServiceProvider services,
+            string applicationName,
+            string environmentName,
+            IList<string> diagnosticMessages)
+        {
+            if (string.IsNullOrEmpty(applicationName))
+            {
+                throw new ArgumentException("Value cannot be null or empty.", "applicationName");
+            }
+
+            var assembly = Assembly.Load(new AssemblyName(applicationName));
+            if (assembly == null)
+            {
+                throw new Exception(String.Format("The assembly '{0}' failed to load.", applicationName));
+            }
+
+            var startupNameWithEnv = "Startup" + environmentName;
+            var startupNameWithoutEnv = "Startup";
+
+            // Check the most likely places first
+            var type =
+                assembly.GetType(startupNameWithEnv) ??
+                assembly.GetType(applicationName + "." + startupNameWithEnv) ??
+                assembly.GetType(startupNameWithoutEnv) ??
+                assembly.GetType(applicationName + "." + startupNameWithoutEnv);
+
+            if (type == null)
+            {
+                // Full scan
+                var definedTypes = assembly.DefinedTypes.ToList();
+
+                var startupType1 = definedTypes.Where(info => info.Name.Equals(startupNameWithEnv, StringComparison.Ordinal));
+                var startupType2 = definedTypes.Where(info => info.Name.Equals(startupNameWithoutEnv, StringComparison.Ordinal));
+
+                var typeInfo = startupType1.Concat(startupType2).FirstOrDefault();
+                if (typeInfo != null)
+                {
+                    type = typeInfo.AsType();
+                }
+            }
+
+            if (type == null)
+            {
+                throw new Exception(String.Format("A type named '{0}' or '{1}' could not be found in assembly '{2}'.",
+                    startupNameWithEnv,
+                    startupNameWithoutEnv,
+                    applicationName));
+            }
+
+            var configureMethod = ApplicationStartup.FindConfigureDelegate(type, environmentName);
+            var servicesMethod = ApplicationStartup.FindConfigureServicesDelegate(type, environmentName);
+
+            object instance = null;
+            if (!configureMethod.MethodInfo.IsStatic || (servicesMethod != null && !servicesMethod.MethodInfo.IsStatic))
+            {
+                instance = ActivatorUtilities.GetServiceOrCreateInstance(services, type);
+            }
+
+            return new ApplicationStartup(configureMethod, servicesMethod, instance);
+        }
+
 
         public static ConfigureDelegate FindConfigureDelegate(Type startupType, string environmentName)
         {
